Add JS detection for momentum scroll event computation (#47995)
Summary:

In D65556000, a bug was introduced that can sometimes fire `onMomentumScrollBegin` directly after `onMomentumScrollEnd`, leading the `ScrollView` to believe it's in the "animating" state. In this state, the `ScrollView` will tell the gesture responder system that it intends to handle touch events, preventing any child components from receiving touches.

Changelog: [Internal]

Reviewed By: andrewdacenko

Differential Revision: D66453988