Improve detached keyboard detection on iOS
Summary:
The iOS keyboard may be in one of three states:
1) floating (previously supported with a width check)
2) split
3) or undocked.

In addition, when using Stage Manager, the keyboard may be wider than
the window itself. This would cause the floating keyboard check to
incorrectly set the bottom position to zero.

Instead, rely on the fact that the UIKeyboardWillHideNotification
notification is sent when the keyboard is in any detached state.

This requires listening for UIKeyboardWillShowNotification instead of
UIKeyboardWillChangeFrameNotification. This is fine, since the show
notification is also sent when the keyboard resizes while open.

Combined with the coordinate system adjustments in the previous commit,
this also fixes an issue with Stage Manager, since the keyboard may be
attached yet wider than the application window.

Changelog:
[iOS] [FIXED] - Improve detached keyboard detection, support Stage Manager on iOS