Skip hidden folders when looking for third party components (#48182)
Summary:

Maintainers from SVG reached out because of an edge case they inencountered when generating the ComponentProvider. In their setup, they had a `.git` folder in the repo and the algorithm was spending a lot of time crawling the git folder.

In general, we should avoid crawling hidden folders.

This change fix that.

## Changelog:
[Codegen][Fixed] - Skip hidden folders when looking for third party components.

Differential Revision: D66959345